move authentication context to contexts folder
