+++ Cinecompass/src/navigation/index.jsx
@@ -3,7 +3,7 @@ import { NavigationContainer } from "@react-navigation/native";
 
 import ScreensNavigator from "./screens.navigator";
 import AccountNavigator from "./account.navigator";
-import { AuthenticationContext } from "../services/authentication.context";
+import { AuthenticationContext } from "../contexts/authentication.context";
 
 const Navigator = () => {
   const { isAuthenticated } = useContext(AuthenticationContext);
